+++ werkbank/environments/CSharpEnvironment.cs
@@ -8,7 +8,7 @@ namespace werkbank.environments
 
         public override string Name => "C#";
         public override string Handle => "csharp";
-        public override string Directory => "code\\csharp";
+        public override string Directory => "code\\dotnet";
 
         public override EnvironmentPreset Preset => new(
             CompressOnArchive: true