+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+[v3.2.0](https://github.com/sedlatschek/werkbank3/releases/tag/v3.2.0.0)
+- change directory of environment C# from code/csharp to code/dotnet
+
 [v3.1.0](https://github.com/sedlatschek/werkbank3/releases/tag/v3.1.0.0)
 - fix bug where the existing instance would not be brought to the front when an additional instance of werkbank was launched
 - add delete button